Add files via upload
In this version, I have added products on the product page. I used JavaScript to pull this information as I find that this is easier than hard coding all the products if I was to hardcode all the products in, it would make it harder for future thinking and if the website and business were to grow then it would make it harder to add more products and slow down the code.