Add files via upload
In this update I have added cart functionality, I have run into some errors like the cart showing up as a null row because it is running through the loop to many times which I will work on fixing. also, the cart number is incorrect and I shall also fix that.