Add files via upload
In this update, I have made it so that it has a more modern look to the product page as I thought that it fit better with the look I am going for. I have also added a popup to the products page for when you click on the add to cart button it brings up a popup that lets you change the quantity and size of a product before you add it to the cart.