Add files via upload
v13. This is the final update which has everything commented on and explained, i have run the code through a validation checker which showed me things I was missing like the alternative text to my images for accessibility and also adding the language to the page. Hence, the browser knows what language it is in.